Fix renaming issues by sticking with 'mach_learning' instead of new 'ml' name
