+++ bcipy/signal/model/demo/mach_learning/demo_cv.py
@@ -1,8 +1,8 @@
 import numpy as np
-from bcipy.signal.model.ml.cross_validation import cross_validation
-from bcipy.signal.model.ml.classifier import RegularizedDiscriminantAnalysis
-from bcipy.signal.model.ml.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
-from bcipy.signal.model.ml.pipeline import Pipeline
+from bcipy.signal.model.mach_learning.cross_validation import cross_validation
+from bcipy.signal.model.mach_learning.classifier import RegularizedDiscriminantAnalysis
+from bcipy.signal.model.mach_learning.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
+from bcipy.signal.model.mach_learning.pipeline import Pipeline
 
 
 def _demo_cv():
