+++ bcipy/signal/model/demo/mach_learning/demo_density_estimation.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from scipy.stats import norm, iqr
-from bcipy.signal.model.ml.density_estimation import KernelDensityEstimate
+from bcipy.signal.model.mach_learning.density_estimation import KernelDensityEstimate
 import matplotlib.pyplot as plt
