+++ bcipy/signal/model/demo/mach_learning/demo_dim_reduction.py
@@ -1,4 +1,4 @@
-from bcipy.signal.model.ml.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
+from bcipy.signal.model.mach_learning.dimensionality_reduction import ChannelWisePrincipalComponentAnalysis
 import numpy as np
